Extract authTokenSchema to avoid Node.js imports in renderer bundle
